update 4x4 to 3x3 matrix, add sep tests file
